# backend/docs/swagger.yaml
basePath: /
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.AuthResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        type: integer
      permissions:
        items:
          type: string
        type: array
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  models.IssueWarningRequest:
    properties:
      deadlineLabel:
        type: string
      letterNumber:
        type: string
      letterType:
        type: string
      reason:
        type: string
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  models.RecordRecipeAuditRequest:
    properties:
      auditor:
        type: string
      avgDeviationPct:
        type: number
      notes:
        type: string
      tkpiStatus:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      fullName:
        type: string
      npsn:
        type: string
      password:
        type: string
      role:
        type: string
      schoolName:
        type: string
      sppgId:
        type: string
    type: object
  models.ReinstateKitchenRequest:
    properties:
      initialQuota:
        type: integer
      reason:
        type: string
    type: object
...
  models.SPPGKitchen:
    properties:
      activeQuota:
        type: integer
      address:
        type: string
      assignedSchools:
        items:
          $ref: '#/definitions/models.SppgAssignedSchool'
        type: array
      avgDeviationPct:
        type: number
      city:
        type: string
      cluster:
        type: string
      code:
        type: string
      coldChainScore:
        type: number
      complianceTrend:
        items:
          type: number
        type: array
      compositeScore:
        type: number
      coordinates:
        type: string
      createdAt:
        type: string
      fleetCount:
        type: integer
      fleetType:
        type: string
      grade:
        type: string
      id:
        type: string
      kitchenArea:
        type: string
      legalEntity:
        type: string
      managerName:
        type: string
      managerNip:
        type: string
      managerPhone:
        type: string
      maxDailyPortions:
        type: integer
      name:
        type: string
      nutritionistName:
        type: string
      nutritionistStr:
        type: string
      province:
        type: string
      quotaUpdatedAt:
        type: string
      quotaUpdatedBy:
        type: string
      recipeAuditedAt:
        type: string
      recipeAuditor:
        type: string
      safetyScore:
        type: number
      staffCount:
        type: integer
      status:
        type: string
      subdistrict:
        type: string
      suspendedAt:
        type: string
      suspendedReason:
        type: string
      timelinessScore:
        type: number
      tkpiStatus:
        description: Status audit & teguran; kolom nullable di DB, dinormalkan di
          sini.
        type: string
      type:
        type: string
      typeLabel:
        type: string
      updatedAt:
        type: string
      warningLetters:
        items:
          $ref: '#/definitions/models.SppgWarningLetter'
        type: array
    type: object
...
  models.SppgAssignedSchool:
    properties:
      city:
        type: string
      name:
        type: string
      npsn:
        type: string
      portionsToday:
        type: integer
    type: object
  models.SppgWarningLetter:
    properties:
      createdAt:
        type: string
      deadlineLabel:
        type: string
      id:
        type: string
      issuedBy:
        type: string
      letterNumber:
        type: string
      reason:
        type: string
      sppgId:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  models.SuspendKitchenRequest:
    properties:
      alternativeSppgId:
        type: string
      reason:
        type: string
    type: object
  models.UpdateQuotaRequest:
    properties:
      quota:
        type: integer
      reason:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      npsn:
        type: string
      role:
        type: string
      schoolName:
        type: string
      sppgId:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://localhost:8080
  description: RESTful API Service built with Golang, Clean Code, and SOLID principles.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: MBG Backend API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Authenticate user and issue tokens
...
  /api/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Return the authenticated user profile and permissions
  /api/auth/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Rotate refresh token and issue a new access token
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: registration data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Create a new validator/siswa account
...
  /sppg:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SPPGKitchen'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Direktori dapur SPPG
      tags:
      - sppg
  /sppg/{id}:
    get:
      parameters:
      - description: SPPG ID
        example: SPPG-01
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SPPGKitchen'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Detail satu dapur SPPG
      tags:
      - sppg
  /sppg/{id}/quota:
    put:
      consumes:
      - application/json
      description: Superadmin saja. Kuota tidak boleh melebihi maxDailyPortions.
      parameters:
      - description: SPPG ID
        in: path
        name: id
        required: true
        type: string
      - description: Kuota baru + alasan
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.UpdateQuotaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SPPGKitchen'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Perbarui kuota produksi
      tags:
      - sppg
  /sppg/{id}/recipe-audit:
    post:
      consumes:
      - application/json
      parameters:
      - description: SPPG ID
        in: path
        name: id
        required: true
        type: string
      - description: Hasil audit
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.RecordRecipeAuditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SPPGKitchen'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Audit gramatur resep TKPI
      tags:
      - sppg
  /sppg/{id}/reinstate:
    post:
      consumes:
      - application/json
      description: Superadmin saja. Mengembalikan status dapur menjadi active dan
        menetapkan kuota awal.
      parameters:
      - description: SPPG ID
        in: path
        name: id
        required: true
        type: string
      - description: Alasan pemulihan + kuota awal
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ReinstateKitchenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SPPGKitchen'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Pulihkan hak operasional dapur SPPG
      tags:
      - sppg
  /sppg/{id}/suspension:
    post:
      consumes:
      - application/json
      description: Superadmin saja. Wajib menunjuk dapur alternatif yang aktif dan
        mampu.
      parameters:
      - description: SPPG ID
        in: path
        name: id
        required: true
        type: string
      - description: Alasan + dapur alternatif
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.SuspendKitchenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SPPGKitchen'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Bekukan dapur SPPG
      tags:
      - sppg
  /sppg/{id}/warnings:
    post:
      consumes:
      - application/json
      description: Superadmin saja. SP-2 hanya sah bila SP-1 sebelumnya sudah terbit.
      parameters:
      - description: SPPG ID
        in: path
        name: id
        required: true
        type: string
      - description: Jenis, nomor, alasan, tenggat
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.IssueWarningRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SPPGKitchen'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Terbitkan surat peringatan SP-1/SP-2
      tags:
      - sppg
...
swagger: "2.0"